Fix sync_file_range() with musl 1.2.4
musl 1.2.4 has removed the transitional LFS off64_t type.
sync_file_range is declared with off_t in musl, which is always 64
bits.

This assumes that the same is true of any other libc which doesn't
provide off64_t.  If it's not, gcc will produce an error due to the
conflicting types of sync_file_range(), so it will be caught and can
be fixed.